+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 0.2 - TODO
 - fix typo in reported exception
+- add basic type description for unknown arguments
 
 ### 0.1 - 2015-11-21
 - Initial release